// projects/sial-paris/src/components/Layout/TopBar/types.ts
export interface Props {
  // sialp:socialLink
  iconClass?: string;
  label?: string;
  "j:linkType"?: string;
  "j:linknode"?: unknown;
  "j:url"?: string;

  // sialp:topBar
  communityText?: string;
  exposantCtaLabel?: string;
  exposantCtaUrl?: string;
  espaceCtaLabel?: string;
  espaceCtaUrl?: string;
}